Java bugs fix

+++ DistributedAuctionSystem/WebSocketClient/src/main/java/it/unipi/dsmt/das/ws/client/WSClient.java
@@ -20,11 +20,11 @@ public class WSClient {
         this(0);
     }
 
-    public WSClient(int auction){
+    public WSClient(long auction){
         this("ws://localhost:8080/DAS/auctions_backend/", auction);
     }
 
-    public WSClient(String endpoint, int auction) {
+    public WSClient(String endpoint, long auction) {
         try {
             URI endpointURI = new URI(endpoint + String.valueOf(auction));
             WebSocketContainer container = ContainerProvider.getWebSocketContainer();
